int rpm -> float rpm

+++ firmware/controllers/algo/advance_map.h
@@ -20,5 +20,5 @@ size_t getMultiSparkCount(int rpm);
 
 class IgnitionState : public ignition_state_s {
 public:
-	floatms_t getSparkDwell(int rpm);
+	floatms_t getSparkDwell(float rpm);
 };
